Fixed Race and BigDecimal in java and sql

+++ services/accounts/src/main/java/com/payflow/accounts/api/dto/CreateAccountRequest.java
@@ -1,5 +1,7 @@
 package com.payflow.accounts.api.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -8,6 +10,6 @@ import jakarta.validation.constraints.PositiveOrZero;
 public record CreateAccountRequest(
         @NotBlank(message = "customerName is required") String customerName,
         @NotBlank @Email(message = "email must be a valid address") String email,
-        @NotNull @PositiveOrZero(message = "openingBalance cannot be negative") Double openingBalance
+        @NotNull @PositiveOrZero(message = "openingBalance cannot be negative") BigDecimal openingBalance
 ) {
 }
